feat: separate notifications to a component
