+++ src/views/mainGridView.jsx
@@ -1,9 +1,7 @@
-import React, { useContext, useEffect } from "react";
+import React, { useContext } from "react";
 import {
   selectPeers,
   useHMSStore,
-  useHMSNotifications,
-  HMSNotificationTypes,
 } from "@100mslive/hms-video-react";
 import { GridCenterView, GridSidePaneView } from "./components/gridView";
 import { AppContext } from "../store/AppContext";
@@ -12,7 +10,6 @@ import { ROLES } from "../common/roles";
 export const MainGridView = ({ isChatOpen, toggleChat, role }) => {
   const { maxTileCount } = useContext(AppContext);
   const peers = useHMSStore(selectPeers);
-  const notification = useHMSNotifications();
   const teacherPeers = peers.filter(
     peer => peer.role.toLowerCase() === ROLES.TEACHER
   );
@@ -25,28 +22,6 @@ export const MainGridView = ({ isChatOpen, toggleChat, role }) => {
 
   const centerPeers = role === ROLES.TEACHER ? studentPeers : teacherPeers;
   const sidebarPeers = role === ROLES.TEACHER ? teacherPeers : studentPeers;
-
-  useEffect(() => {
-    if (!notification) {
-      return;
-    }
-    switch (notification.type) {
-      case HMSNotificationTypes.PEER_JOINED:
-        console.log("[Peer Joined]", notification);
-        break;
-      case HMSNotificationTypes.PEER_LEFT:
-        console.log("[Peer Left]", notification);
-        break;
-      case HMSNotificationTypes.NEW_MESSAGE:
-        console.log("[New Message]", notification);
-        break;
-      case HMSNotificationTypes.ERROR:
-        console.log("[Error]", notification);
-        break;
-      default:
-        break;
-    }
-  }, [notification]);
 
   return (
     <React.Fragment>
